refactor(css): split app.css into game-header, configs, reports, modals, home (#67)
Original app.css (2128 lines) split into 5 files under css/app/:
- game-header.css (774) - search, game header, deck status, external links
- configs.css (121) - config cards section
- reports.css (740) - report cards, author, signals, votes, flag modal
- modals.css (290) - login modal, gist bar, edit report modal
- home.css (239) - filter bar, view controls, scoring guide, stub page

All HTML pages using app.css updated. No CSS rules changed.
