+++ js/app/main.js
@@ -1,7 +1,7 @@
 // Entry point for the app page: bootstraps routing and search wiring.
 // (Replaces the inline bootstrap that lived at the top/bottom of app.js.)
-import { route } from './router.js?v=3654cd05';
-import { wireSearch } from './components/search.js?v=e2d2ab29';
+import { route } from './router.js?v=dded2c1b';
+import { wireSearch } from './components/search.js?v=a7e41fb5';
 
 window.addEventListener('hashchange', () => route());
 window.addEventListener('popstate', () => route());
