+++ submit.html
@@ -12,7 +12,11 @@
 <link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
 <link rel="stylesheet" href="css/shared/site.css?v=86052a10">
 <link rel="stylesheet" href="css/shared/cards.css?v=c0a5fad1">
-<link rel="stylesheet" href="css/app/app.css?v=c7bfd836">
+<link rel="stylesheet" href="css/app/game-header.css?v=cec66a65">
+<link rel="stylesheet" href="css/app/configs.css?v=90d5bed5">
+<link rel="stylesheet" href="css/app/reports.css?v=bb420673">
+<link rel="stylesheet" href="css/app/modals.css?v=68ffdfa5">
+<link rel="stylesheet" href="css/app/home.css?v=13fd2117">
 </head>
 <body>
